pyproject.toml
[project]
name = "dragon-ml-toolbox"
version = "20.0.0"
description = "Complete pipelines and helper tools for data science and machine learning projects."
authors = [
    { name = "Karl Luigi Loza Vidaurre", email = "luigiloza@gmail.com" }
]
readme = "README.md"
requires-python = "==3.12.*"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = []

[project.urls]
Homepage = "https://github.com/DrAg0n-BoRn/ML_tools"
Changelog = "https://github.com/DrAg0n-BoRn/ML_tools/blob/master/CHANGELOG.md"

[project.optional-dependencies]
# Machine Learning main toolbox. Additionally Requires PyTorch and torchvision with CUDA support
ML = [
    # Core
    "torch",
    "torchvision",
    "numpy<2.0",          # Pinned for PyTorch extension compatibility
    "pandas",
    "polars>=1.0",
    "pyarrow",            # Essential for Polars efficiency
    "joblib",
    "tqdm",
    "colorlog",
    
    # Math / ML
    "scikit-learn",
    "shap",               # Uses KernelExplainer/DeepExplainer
    "captum",
    "evotorch",
    "torchmetrics",

    # Viz / Notebook
    "matplotlib",
    "seaborn",
    "plotly",              # Interactive plots
    "Pillow",
    "ipython",
    "ipykernel",
    "notebook",
    "jupyterlab",
    "ipywidgets",
]

# ensemble tools
ensemble = [
    # Core
    # "numpy>=2.0",         # Root dependency
    "numpy",               # should be >2.0
    "numba>=0.60",        # Required for NumPy 2.0 JIT compilation
    "shap>=0.46",         # Required to use the newer Numba
    
    # Data
    "pandas",
    "polars>=1.0",        # Recommended for safest NumPy 2.0
    "pyarrow",
    "joblib",
    "tqdm",
    "colorlog",

    # Math / ML
    "scikit-learn>=1.5,<1.8",  # Required >1.5 for numpy 2.0 but less than 1.8 for compatibility with imbalanced-learn
    "imbalanced-learn",
    "xgboost",
    "lightgbm",

    # Viz / Notebook
    "matplotlib",
    "seaborn",
    "ipython",
    "ipykernel",
    "notebook",
    "jupyterlab",
    "ipywidgets",
]

# MICE and VIF
mice = [
    "numpy<2.0",
    "pandas",
    "polars>=1.0",
    "joblib",
    "miceforest>=6.0.0",
    "plotnine>=0.12",
    "matplotlib",
    "statsmodels",
    "lightgbm<=4.5.0",
    "shap",
    "colorlog", 
    "pyarrow"
]

# Process excel files
excel = [
    "pandas",
    "openpyxl",
    "ipython",
    "ipykernel",
    "notebook",
    "jupyterlab",
    "ipywidgets",
    "colorlog"
]

# GUI for boost algorithms
# Optional: matplotlib / seaborn
gui-boost = [
    # "numpy>=2.0",
    "numpy", # should be >2.0
    "joblib",
    "FreeSimpleGUI>=5.2",
    "xgboost", 
    "lightgbm", 
]

# GUI for Pytorch 
# Optional: Pillow / Torchvision / matplotlib / seaborn
gui-torch = [
    "numpy<2.0",
    "torch",
    "FreeSimpleGUI>=5.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# [tool.setuptools]
# include = ["ml_tools"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ml_tools*"]
namespaces = false

[dependency-groups]
dev = [
    "captum>=0.8.0",
    "colorlog>=6.10.1",
    "evotorch>=0.6.1",
    "freesimplegui>=5.2",
    "ipykernel>=7.1.0",
    "ipython>=9.8.0",
    "ipywidgets>=8.1.8",
    "joblib>=1.5.2",
    "jupyterlab>=4.5.0",
    "matplotlib>=3.10.8",
    "notebook>=7.5.0",
    "numpy<2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "polars>=1.0",
    "pyarrow>=22.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "torch>=2.9.1",
    "torchmetrics>=1.5.2",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
